Fixed the bug
It was an issue with the email addresses config file.
I thought it would be overwritten by a git pull but it wasn't, so it
contained the email addresses repeatedly...
I've wiped the config file now.
I've set the code up to only send once to an address, whether successful
or not.